// web/src/lib/types.ts
export type FileType = "media" | "photo" | "video" | "audio" | "file";
export type DownloadStatus =
  | "idle"
  | "downloading"
  | "paused"
  | "completed"
  | "error";

export type TransferStatus = "idle" | "transferring" | "completed" | "error";

export type TelegramFile = {
  id: number;
  telegramId: number;
  uniqueId: string;
  messageId: number;
  chatId: number;
  fileName: string;
  type: FileType;
  mimeType?: string;
  size: number;
  downloadedSize: number;
  thumbnail?: string;
  thumbnailFile?: Thumbnail;
  downloadStatus: DownloadStatus;
  date: number;
  formatDate: string;
  caption: string;
  localPath: string;
  hasSensitiveContent: boolean;
  startDate: number;
  completionDate: number;
  originalDeleted: boolean;
  transferStatus?: TransferStatus;
  extra?: PhotoExtra | VideoExtra;
  tags?: string;
  loaded: boolean;
  threadChatId: number;
  messageThreadId: number;
  hasReply?: boolean;
  reactionCount: number;

  prev?: TelegramFile;
  next?: TelegramFile;
};

export type PhotoExtra = {
  width: number;
  height: number;
  type: string;
};

export type VideoExtra = {
  width: number;
  height: number;
  duration: number;
  mimeType: string;
};

export type Thumbnail = {
  uniqueId: string;
  mimeType: string;
  extra: {
    width: number;
    height: number;
  };
};
// ...
export type SortFields = "date" | "completion_date" | "size" | "reaction_count";

export type FileFilter = {
  search: string;
  type: FileType | "all";
  downloadStatus?: DownloadStatus;
  transferStatus?: TransferStatus;
  offline: boolean;
  tags: string[];
  dateType?: "sent" | "downloaded";
  dateRange?: [string, string];
  sizeRange?: [number, number];
  sizeUnit?: "KB" | "MB" | "GB";
  sort?: SortFields;
  order?: "asc" | "desc";
};